Fix abort propagation in podcast fetches
